Fix Esphome failure to compile
"Custom" components have been deprecated in version 2025.2
This is a "custom" component that is too complex to convert to an "external" component on my own without MUCH better documentation. (If it is even possible within the constraints of Esphome core...)

Changes: Add an "external" component to bring back the function of "custom" components, and allow successful compile.